+++ classes/com/sapienter/jbilling/server/util/IWebServicesSessionBean.java
@@ -20,6 +20,7 @@
 
 package com.sapienter.jbilling.server.util;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.jws.WebService;
@@ -335,7 +336,7 @@ public interface IWebServicesSessionBean {
     public Integer[] getLastOrdersByItemType(Integer userId, Integer itemTypeId,
             Integer number) throws SessionInternalError;
 
-	public Double isUserSubscribedTo(Integer userId, Integer itemId);
+	public BigDecimal isUserSubscribedTo(Integer userId, Integer itemId);
 
 	public Integer[] getUserItemsByCategory(Integer userId, Integer categoryId);
